requirements.txt
tensorflow-gpu==1.14.0
pandas
matplotlib